chore(tests): split up e2e tests - portal maker
