+++ .github/workflows/test.yml
@@ -426,7 +426,22 @@ jobs:
             'dashboard/**/*.spec.js',
             'feedback/**/*.spec.js',
             'login/**/*.spec.js',
-            'maker/**/*.spec.js',
+            'maker/check-deal-details/*.spec.js',
+            'maker/clone-deal/*.spec.js',
+            'maker/dashboard/*.spec.js',
+            'maker/fill-and-submit-issue-facility-form/*.spec.js',
+            'maker/submit-a-deal-for-review/*.spec.js',
+            'maker/submit-a-deal-for-review-with-issued-and-unissued-facilities-and-resubmit-after-checker-returns/*.spec.js',
+            'maker/submit-issued-facilities-for-review/*.spec.js',
+            'maker/bank-details/*.spec.js',
+            'maker/eligibility/*.spec.js',
+            'maker/abandon-a-deal/*.spec.js',
+            'maker/before-you-start/*.spec.js',
+            'maker/bss-mandatory-criteria/*.spec.js',
+            'maker/edit-deal-name/*.spec.js',
+            'maker/proceed-to-review/*.spec.js',
+            'maker/select-scheme/*.spec.js',
+            'maker/xss/*.spec.js',
             'maker-about-supply-contract/**/*.spec.js',
             'maker-bond/**/*.spec.js',
             'maker-facilities/**/*.spec.js',
