+++ e2e-tests/portal/cypress/e2e/journeys/maker/bank-details/bank-details-created-field.spec.js
@@ -1,6 +1,6 @@
-const { bankDetails } = require('../../pages');
-const relative = require('../../relativeURL');
-const MOCK_USERS = require('../../../../../e2e-fixtures');
+const { bankDetails } = require('../../../pages');
+const relative = require('../../../relativeURL');
+const MOCK_USERS = require('../../../../../../e2e-fixtures');
 
 const { BANK1_MAKER1 } = MOCK_USERS;
