+++ e2e-tests/portal/cypress/e2e/journeys/maker/bank-details/bank-details.spec.js
@@ -1,7 +1,7 @@
-const { bankDetails, contract, contractCheckDealDetails } = require('../../pages');
-const partials = require('../../partials');
-const relative = require('../../relativeURL');
-const MOCK_USERS = require('../../../../../e2e-fixtures');
+const { bankDetails, contract, contractCheckDealDetails } = require('../../../pages');
+const partials = require('../../../partials');
+const relative = require('../../../relativeURL');
+const MOCK_USERS = require('../../../../../../e2e-fixtures');
 
 const { BANK1_MAKER1 } = MOCK_USERS;
