+++ e2e-tests/portal/cypress/e2e/journeys/maker/bss-mandatory-criteria/bss-mandatory-criteria.spec.js
@@ -1,8 +1,8 @@
-const pages = require('../../pages');
-const partials = require('../../partials');
-const dealFullyCompleted = require('./fixtures/dealFullyCompleted');
-const relative = require('../../relativeURL');
-const MOCK_USERS = require('../../../../../e2e-fixtures');
+const pages = require('../../../pages');
+const partials = require('../../../partials');
+const dealFullyCompleted = require('../fixtures/dealFullyCompleted');
+const relative = require('../../../relativeURL');
+const MOCK_USERS = require('../../../../../../e2e-fixtures');
 
 const { BANK1_MAKER1 } = MOCK_USERS;
